+++ assets/javascripts/kitten/components/meters/progress-ring/index.js
@@ -6,9 +6,8 @@ import { pxToRem } from '../../../helpers/utils/typography'
 
 const getDashLength = radius => 2 * Math.PI * radius
 
-const getDashOffset = ({ r, progressValue }) => (
-  getDashLength(r) * (100 - progressValue) / 100
-)
+const getDashOffset = ({ r, progressValue }) =>
+  (getDashLength(r) * (100 - progressValue)) / 100
 
 const rotateAnimate = ({ r, progressValue }) => keyframes`
   from {
@@ -105,7 +104,7 @@ ProgressRing.propTypes = {
     Radius value of ProgressRing
   */
   radius: PropTypes.number,
-   /**
+  /**
     Width of circle (stroke)
   */
   strokeWidth: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),